requirements.txt
openai-whisper
sounddevice
numpy
rumps
pynput
pyperclip
